# Documentation/API/openapi_core.yaml
openapi: 3.0.3
info:
  title: E.V.A. Core API
  description: |
    API principale d'E.V.A. - L'Orchestrateur Central (Expert A).
    Gère le routage des requêtes, la mémoire conversationnelle et la synthèse des réponses.
  version: 1.0.0
  contact:
    name: The Hive Admin
servers:
  - url: http://the-hive.local:8000/api/v1
    description: Serveur local (via Tailscale)
...
paths:
...
  /memory/search:
    post:
      tags: [Memory]
      summary: Recherche sémantique dans la mémoire
      operationId: searchMemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySearchRequest'
      responses:
        '200':
          description: Résultats de recherche
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySearchResponse'
...
components:
  schemas:
...
    MemorySearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Requête de recherche sémantique
        collection:
          type: string
          enum: [conversations, osint, trading, general]
          default: general
        limit:
          type: integer
          default: 5
          maximum: 20
        min_score:
          type: number
          minimum: 0
          maximum: 1
          default: 0.7
    MemorySearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/MemoryItem'
        query_time_ms:
          type: integer
    MemoryItem:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        score:
          type: number
        metadata:
          type: object
          properties:
            source:
              type: string
            timestamp:
              type: string
              format: date-time
            tags:
              type: array
              items:
                type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via authentification Tailscale + Voix
security:
  - BearerAuth: []